requirements.txt
scipy==1.1.0
boto==2.49.0
matplotlib==2.2.3
ipdb==0.11
pandas==0.23.4
seaborn==0.9.0
numpy==1.15.1
scikit_learn==0.20.2
PyYAML==3.13
docopt==0.6.2
